remove NowPlaying, consolidate into one file
