+++ src/views/Jukebox/jukebox.vue
@@ -7,7 +7,32 @@
         <div class="inner-background">
           <h1 class="title"><font-awesome-icon class="icon" :icon="icons.faRadio" /> Meet Jukebox</h1>
           <p class="subtitle">Discover original music from Stevenson students.</p>
-          <NowPlaying :song="currentSong" />
+          <div class="now-playing">
+            <img class="album-art" v-if="currentSong.album != 'NO_ART'" :src="`/music/album/${currentSong.album}`" />
+            <div class="song-info">
+              <p class="song-title">{{currentSong.name}}</p>
+              <p class="song-bullet">&bull;</p>
+              <p class="song-artist">{{currentSong.artist}}, {{currentSong.year}}</p>
+            </div>
+            <div class="controls">
+              <button class="control-button">
+                <font-awesome-icon :icon="icons.faStepBackward" />
+              </button>
+              <button @click="playing ? pause() : play(); playing = !playing" class="control-button play-button">
+                <font-awesome-icon :icon="playing ? icons.faPause : icons.faPlay" />
+              </button>
+              <button class="control-button">
+                <font-awesome-icon :icon="icons.faStepForward" />
+              </button>
+            </div>
+            <div class="progress">
+              <p class="progress-text">{{formatTime(currentTimePlaying)}}</p>
+              <div class="progress-bar">
+                <div class="progress-bar-fill" :style="`width: ${percentFinished}%;`" />
+              </div>
+              <p class="progress-text">{{formatTime(totalSongTime)}}</p>
+            </div>
+          </div>
           <div class="song-list" v-for="(song, index) in songs" :key="index">
             <button @click="setSongOf(index)" class="play-button">
               <font-awesome-icon :icon="icons.faPlay" />
@@ -23,30 +48,64 @@
 
 <script>
 import HomeLink from '@/components/HomeLink.vue';
-import NowPlaying from '@/components/NowPlaying.vue';
 import songs from '@/data/music.json';
 import {
   faRadio,
   faPlay,
+  faStepBackward,
+  faStepForward,
+  faPause,
 } from '@fortawesome/free-solid-svg-icons';
+import { Howl } from 'howler';
 
 export default {
-  components: { HomeLink, NowPlaying },
+  components: { HomeLink },
   data() {
     return {
       icons: {
-        faRadio, faPlay,
+        faRadio, faPlay, faStepBackward, faStepForward, faPause,
       },
       currentSong: songs[0],
       songs,
     };
   },
   mounted() {
-    console.log(this.songs[0]);
+    const fileFormatRegex = /\.[0-9a-z]+$/i;
+    for (const song of songs) {
+      song.howler = new Howl({
+        src: [`/music/${song.file}`],
+        format: [song.file.match(fileFormatRegex)[0]],
+        volume: 1,
+      });
+    }
   },
   methods: {
     setSongOf(index) {
       this.currentSong = this.songs[index];
+    },
+    play() {
+      this.currentSong.howler.play();
+      setInterval(() => this.updateProgressBar(), 100);
+      this.currentSong.howler.fade(0, 1, 75);
+    },
+    async pause() {
+      this.currentSong.howler.fade(1, 0, 200, this.songId);
+      await this.sleep(200);
+      this.currentSong.howler.pause();
+    },
+    updateProgressBar() {
+      this.percentFinished = (this.currentSong.howler.seek() / this.currentSong.howler.duration()) * 100;
+      this.currentTimePlaying = this.currentSong.howler.seek();
+      this.totalSongTime = this.currentSong.howler.duration();
+    },
+    formatTime(seconds) {
+      const minutes = Math.floor(seconds / 60);
+      const secondsLeft = (seconds - (minutes * 60)).toFixed(0); // ensure seconds is a whole number (0 decimal places)
+      return `${minutes}:${seconds < 10 ? `0${secondsLeft}` : secondsLeft}`;
+    },
+    sleep(ms) {
+      // eslint-disable-next-line no-promise-executor-return
+      return new Promise((resolve) => setTimeout(resolve, ms));
     },
   },
 };
@@ -101,6 +160,82 @@ export default {
       top: 30px
       +mobile
         top: 20px
+
+    .now-playing
+      .progress
+        display: flex
+        .progress-text
+          font-size: 0.8em
+          font-weight: bold
+          color: var(--textColor)
+        .progress-bar
+          width: 80%
+          display: flex
+          margin: 15px auto auto auto
+          positon: absolute
+          text-align: center
+          background-color: var(--tertiary)
+          border-radius: 10px
+        .progress-bar-fill
+          height: 15px
+          width: 20%
+          background-color: var(--headerBackgroundColor)
+          transition: width 0.5s ease-in-out
+          border-radius: 10px
+
+      .album-art
+        display: block
+        margin-top: 0px
+        width: 300px
+        height: 300px
+        object-fit: cover
+        margin-left: auto
+        margin-right: auto
+        border-radius: 10px
+
+      .song-info
+        display: flex
+        justify-content: center
+
+        .song-title
+          font-size: 1.5em
+          font-weight: bold
+        .song-bullet
+          font-weight: bold
+          font-size: 1.55em
+          margin: auto 10px auto 10px
+          color: var(--secondary)
+        .song-artist
+          font-size: 1.5em
+          font-weight: bold
+          color: var(--secondary)
+
+      .controls
+        display: flex
+        justify-content: center
+
+        .control-button
+          background-color: transparent
+          border: none
+          margin: 0 10px 0 10px
+          font-size: 2em
+          color: var(--secondary)
+          &:hover
+            color: var(--headerBackgroundColor)
+            cursor: pointer
+        .play-button
+          background-color: var(--headerBackgroundColor)
+          border-radius: 50%
+          padding: 10px 20px 9px 21px
+          margin: 0
+          transition: 0.15s
+          font-size: 2.5em
+          color: white
+          &:hover
+            color: white
+            font-size: 2.6em
+            cursor: pointer
+            box-shadow: 0 0 10px 0 var(--tertiary)
 
     .song-list
       display: flex
